[ALPS08148460] GPU: GPUSWERRATA-1559(c5678/2)
[Description]
This commit introduces a set of pre-allocated MCU shared VA regions
for runtime mapping of on-slot groups suspend buffers and their
bound queues user I/O pages. The pre-allocated set is owned by the
scheduler, which lazily binds resources to a group only when it is
assigned to a GPU slot.

The MMU mapping in these pre-allocated MCU shared VA regions, once
bound to a group and mapped, is lazily unmapped only when the resource
is required by other groups. For the sake of performance, the size
of the pre-allocated set is greater than strictly necessary and
configured to be a multiple of the number of GPU slots.

The purpose of this feature is mitigating potential DoS attacks aimed
at exhausting the GPUs limited MCU shared VA memory by creating a large
number of groups.

[Patch]
GPUCORE-35854 Mapping MCU Shared Regions At Runtime

[ARM Ticket]
03400817

MTK-Commit-Id: 3dc168315f4bf0607725c279fc0e3fa68002757f

CR-Id: ALPS08148460
Feature: ARM GPU DDK driver
Change-Id: I6d825dbe092e33dc6dc4ed3ed605d2d34abb3762
(cherry-pick from 1da179e766a2bb32b3b48a246d8f2d03e4650444)

Change-Id: I71309c7a2edfef8a150d4038fa6ba79449186821